webapp/requirements.txt
flask
flask-cors
scikit-learn
joblib
gunicorn
numpy
Pillow
